Mms: Add a function that copy SMS into ICC card
Add a context menu into message item to support copy SMS into ICC
card for both subscriptions.

+++ src/com/android/mms/ui/ManageSimMessages.java
@@ -35,8 +35,14 @@ import android.database.sqlite.SqliteWrapper;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Handler;
+import android.provider.ContactsContract.Contacts;
+import android.provider.ContactsContract.CommonDataKinds.Email;
 import android.provider.Telephony.Sms;
 import android.telephony.SmsManager;
+import android.text.TextUtils;
+import android.text.SpannableString;
+import android.text.style.URLSpan;
+import android.text.util.Linkify;
 import android.util.Log;
 import android.view.ContextMenu;
 import android.view.Menu;
@@ -51,7 +57,10 @@ import com.android.internal.telephony.IccCardConstants;
 import com.android.internal.telephony.TelephonyIntents;
 import com.android.mms.LogTag;
 import com.android.mms.R;
+import com.android.mms.data.Contact;
 import com.android.mms.transaction.MessagingNotification;
+
+import java.util.ArrayList;
 
 /**
  * Displays a list of the SMS messages stored on the ICC.
@@ -59,45 +68,52 @@ import com.android.mms.transaction.MessagingNotification;
 public class ManageSimMessages extends Activity
         implements View.OnCreateContextMenuListener {
     private static final Uri ICC_URI = Uri.parse("content://sms/icc");
+    private static final Uri ICC1_URI = Uri.parse("content://sms/icc1");
+    private static final Uri ICC2_URI = Uri.parse("content://sms/icc2");
     private static final String TAG = LogTag.TAG;
     private static final int MENU_COPY_TO_PHONE_MEMORY = 0;
     private static final int MENU_DELETE_FROM_SIM = 1;
     private static final int MENU_VIEW = 2;
+    private static final int MENU_REPLY = 3;
+    private static final int MENU_FORWARD = 4;
+    private static final int MENU_CALL_BACK = 5;
+    private static final int MENU_ADD_ADDRESS_TO_CONTACTS = 6;
+    private static final int MENU_SEND_EMAIL = 7;
     private static final int OPTION_MENU_DELETE_ALL = 0;
 
     private static final int SHOW_LIST = 0;
     private static final int SHOW_EMPTY = 1;
     private static final int SHOW_BUSY = 2;
     private int mState;
+    private long mSubscription;
 
-
+    private Uri mIccUri;
     private ContentResolver mContentResolver;
     private Cursor mCursor = null;
     private ListView mSimList;
     private TextView mMessage;
     private MessageListAdapter mListAdapter = null;
     private AsyncQueryHandler mQueryHandler = null;
+    private boolean mIsDeleteAll = false;
+    private boolean mIsQuery = false;
 
     public static final int SIM_FULL_NOTIFICATION_ID = 234;
-
-    protected BroadcastReceiver mBroadcastReceiver = new BroadcastReceiver() {
-        public void onReceive(Context context, Intent intent) {
-            if (TelephonyIntents.ACTION_SIM_STATE_CHANGED.equals(intent.getAction())) {
-                String stateExtra = intent.getStringExtra(IccCardConstants.INTENT_KEY_ICC_STATE);
-                if (stateExtra != null
-                        && ((IccCardConstants.INTENT_VALUE_ICC_ABSENT.equals(stateExtra)
-                        || IccCardConstants.INTENT_VALUE_ICC_UNKNOWN.equals(stateExtra)))) {
-                    updateState(SHOW_EMPTY);
-                }
-            }
-        }
-    };
 
     private final ContentObserver simChangeObserver =
             new ContentObserver(new Handler()) {
         @Override
         public void onChange(boolean selfUpdate) {
             refreshMessageList();
+        }
+    };
+
+    private final BroadcastReceiver mReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            String action = intent.getAction();
+            if (TelephonyIntents.ACTION_SIM_STATE_CHANGED.equals(action)) {
+                refreshMessageList();
+            }
         }
     };
 
@@ -115,6 +131,10 @@ public class ManageSimMessages extends Activity
         ActionBar actionBar = getActionBar();
         actionBar.setDisplayHomeAsUpEnabled(true);
 
+        IntentFilter filter = new IntentFilter();
+        filter.addAction(TelephonyIntents.ACTION_SIM_STATE_CHANGED);
+        registerReceiver(mReceiver, filter);
+
         init();
     }
 
@@ -128,6 +148,9 @@ public class ManageSimMessages extends Activity
     private void init() {
         MessagingNotification.cancelNotification(getApplicationContext(),
                 SIM_FULL_NOTIFICATION_ID);
+        mSubscription = getIntent().getLongExtra(MessageUtils.SUBSCRIPTION_KEY,
+                MessageUtils.SUB_INVALID);
+        mIccUri = MessageUtils.getIccUriBySubscription(mSubscription);
 
         updateState(SHOW_BUSY);
         startQuery();
@@ -159,6 +182,15 @@ public class ManageSimMessages extends Activity
                             mParent, mCursor, mSimList, false, null);
                     mSimList.setAdapter(mListAdapter);
                     mSimList.setOnCreateContextMenuListener(mParent);
+                    mSimList.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+                        @Override
+                        public void onItemClick(AdapterView<?> parent, View view,
+                                                int position, long id) {
+                            if (view != null) {
+                                ((MessageListItem) view).onMessageListItemClick();
+                            }
+                        }
+                    });
                     updateState(SHOW_LIST);
                 } else {
                     mListAdapter.changeCursor(mCursor);
@@ -171,12 +203,17 @@ public class ManageSimMessages extends Activity
             }
             // Show option menu when query complete.
             invalidateOptionsMenu();
+            mIsQuery = false;
         }
     }
 
     private void startQuery() {
         try {
-            mQueryHandler.startQuery(0, null, ICC_URI, null, null, null, null);
+            if (mIsQuery) {
+                return;
+            }
+            mIsQuery = true;
+            mQueryHandler.startQuery(0, null, mIccUri, null, null, null, null);
         } catch (SQLiteException e) {
             SqliteWrapper.checkSQLiteException(this, e);
         }
@@ -199,8 +236,109 @@ public class ManageSimMessages extends Activity
                  R.string.sim_copy_to_phone_memory);
         menu.add(0, MENU_DELETE_FROM_SIM, 0, R.string.sim_delete);
 
+        Cursor cursor = mListAdapter.getCursor();
+        if (isIncomingMessage(cursor)) {
+            String address = cursor.getString(cursor.getColumnIndexOrThrow("address"));
+            Intent intent = new Intent(Intent.ACTION_SENDTO, Uri.fromParts("smsto", address, null));
+            menu.add(0, MENU_REPLY, 0, R.string.menu_reply).setIntent(intent);
+        }
+        menu.add(0, MENU_FORWARD, 0, R.string.menu_forward);
+        addCallAndContactMenuItems(menu, cursor);
+
         // TODO: Enable this once viewMessage is written.
         // menu.add(0, MENU_VIEW, 0, R.string.sim_view);
+    }
+
+    // Refs to ComposeMessageActivity.java
+    private final void addCallAndContactMenuItems(
+            ContextMenu menu, Cursor cursor) {
+        String address = cursor.getString(cursor.getColumnIndexOrThrow("address"));
+        String body = cursor.getString(cursor.getColumnIndexOrThrow("body"));
+        // Add all possible links in the address & message
+        StringBuilder textToSpannify = new StringBuilder();
+        if (address != null) {
+            textToSpannify.append(address + ": ");
+        }
+        textToSpannify.append(body);
+
+        SpannableString msg = new SpannableString(textToSpannify.toString());
+        Linkify.addLinks(msg, Linkify.ALL);
+        ArrayList<String> uris =
+            MessageUtils.extractUris(msg.getSpans(0, msg.length(), URLSpan.class));
+
+        while (uris.size() > 0) {
+            String uriString = uris.remove(0);
+            // Remove any dupes so they don't get added to the menu multiple times
+            while (uris.contains(uriString)) {
+                uris.remove(uriString);
+            }
+
+            int sep = uriString.indexOf(":");
+            String prefix = null;
+            if (sep >= 0) {
+                prefix = uriString.substring(0, sep);
+                uriString = uriString.substring(sep + 1);
+            }
+            boolean addToContacts = false;
+            if ("mailto".equalsIgnoreCase(prefix))  {
+                String sendEmailString = getString(
+                        R.string.menu_send_email).replace("%s", uriString);
+                Intent intent = new Intent(Intent.ACTION_VIEW,
+                        Uri.parse("mailto:" + uriString));
+                intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_WHEN_TASK_RESET);
+                menu.add(0, MENU_SEND_EMAIL, 0, sendEmailString)
+                    .setIntent(intent);
+                addToContacts = !haveEmailContact(uriString);
+            } else if ("tel".equalsIgnoreCase(prefix)) {
+                String callBackString = getString(
+                        R.string.menu_call_back).replace("%s", uriString);
+                Intent intent = new Intent(Intent.ACTION_CALL,
+                        Uri.parse("tel:" + uriString));
+                intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_WHEN_TASK_RESET);
+                addToContacts = !isNumberInContacts(uriString);
+                if (addToContacts) {
+                    menu.add(0, MENU_CALL_BACK, 0, callBackString).setIntent(intent);
+                } else {
+                    // Get the name by the number if the number has been saved
+                    // in Contacts.
+                    String name = Contact.get(uriString, false).getName();
+                    menu.add(0, MENU_CALL_BACK, 0,
+                            getString(R.string.menu_call_back).replace("%s", name)).setIntent(
+                            intent);
+                }
+            }
+            if (addToContacts) {
+                Intent intent = ConversationList.createAddContactIntent(uriString);
+                String addContactString = getString(
+                        R.string.menu_add_address_to_contacts).replace("%s", uriString);
+                menu.add(0, MENU_ADD_ADDRESS_TO_CONTACTS, 0, addContactString)
+                    .setIntent(intent);
+            }
+        }
+    }
+
+    private boolean haveEmailContact(String emailAddress) {
+        Cursor cursor = SqliteWrapper.query(this, getContentResolver(),
+                Uri.withAppendedPath(Email.CONTENT_LOOKUP_URI, Uri.encode(emailAddress)),
+                new String[] { Contacts.DISPLAY_NAME }, null, null, null);
+
+        if (cursor != null) {
+            try {
+                while (cursor.moveToNext()) {
+                    String name = cursor.getString(0);
+                    if (!TextUtils.isEmpty(name)) {
+                        return true;
+                    }
+                }
+            } finally {
+                cursor.close();
+            }
+        }
+        return false;
+    }
+
+    private boolean isNumberInContacts(String phoneNumber) {
+        return Contact.get(phoneNumber, true).existsInDatabase();
     }
 
     @Override
@@ -225,14 +363,35 @@ public class ManageSimMessages extends Activity
                         updateState(SHOW_BUSY);
                         deleteFromSim(cursor);
                         dialog.dismiss();
+                        invalidateOptionsMenu();
+                        runOnUiThread(new Runnable() {
+                            @Override
+                            public void run() {
+                                refreshMessageList();
+                            }
+                        });
                     }
                 }, R.string.confirm_delete_SIM_message);
                 return true;
+            case MENU_FORWARD: {
+                String smsBody = cursor.getString(cursor.getColumnIndexOrThrow("body"));
+                forwardMessage(smsBody);
+                return true;
+            }
             case MENU_VIEW:
                 viewMessage(cursor);
                 return true;
         }
         return super.onContextItemSelected(item);
+    }
+
+    private void forwardMessage(String smsBody) {
+        Intent intent = new Intent(this, ComposeMessageActivity.class);
+        intent.putExtra("exit_on_sent", true);
+        intent.putExtra("forwarded_message", true);
+        intent.putExtra("sms_body", smsBody);
+        intent.setClassName(this, "com.android.mms.ui.ForwardMessageActivity");
+        startActivity(intent);
     }
 
     @Override
@@ -244,16 +403,27 @@ public class ManageSimMessages extends Activity
     @Override
     public void onPause() {
         super.onPause();
-        unregisterReceiver(mBroadcastReceiver);
         mContentResolver.unregisterContentObserver(simChangeObserver);
+    }
+
+    @Override
+    public void onBackPressed() {
+        if (mIsDeleteAll) {
+            mIsDeleteAll = false;
+        } else {
+            super.onBackPressed();
+        }
+    }
+
+    @Override
+    public void onDestroy() {
+        unregisterReceiver(mReceiver);
+        super.onDestroy();
     }
 
     private void registerSimChangeObserver() {
         mContentResolver.registerContentObserver(
-                ICC_URI, true, simChangeObserver);
-        final IntentFilter intentFilter =
-                new IntentFilter(TelephonyIntents.ACTION_SIM_STATE_CHANGED);
-        registerReceiver(mBroadcastReceiver, intentFilter);
+                mIccUri, true, simChangeObserver);
     }
 
     private void copyToPhoneMemory(Cursor cursor) {
@@ -261,12 +431,13 @@ public class ManageSimMessages extends Activity
                 cursor.getColumnIndexOrThrow("address"));
         String body = cursor.getString(cursor.getColumnIndexOrThrow("body"));
         Long date = cursor.getLong(cursor.getColumnIndexOrThrow("date"));
-
+        int subscription = cursor.getInt(cursor.getColumnIndexOrThrow("sub_id"));
         try {
             if (isIncomingMessage(cursor)) {
-                Sms.Inbox.addMessage(mContentResolver, address, body, null, date, true /* read */);
+                Sms.Inbox.addMessage(subscription, mContentResolver, address, body, null,
+                        date, true /* read */);
             } else {
-                Sms.Sent.addMessage(mContentResolver, address, body, null, date);
+                Sms.Sent.addMessage(subscription, mContentResolver, address, body, null, date);
             }
         } catch (SQLiteException e) {
             SqliteWrapper.checkSQLiteException(this, e);
@@ -284,24 +455,37 @@ public class ManageSimMessages extends Activity
     private void deleteFromSim(Cursor cursor) {
         String messageIndexString =
                 cursor.getString(cursor.getColumnIndexOrThrow("index_on_icc"));
-        Uri simUri = ICC_URI.buildUpon().appendPath(messageIndexString).build();
+        Uri simUri = mIccUri.buildUpon().appendPath(messageIndexString).build();
 
         SqliteWrapper.delete(this, mContentResolver, simUri, null, null);
     }
 
     private void deleteAllFromSim() {
+        mIsDeleteAll = true;
         Cursor cursor = (Cursor) mListAdapter.getCursor();
 
         if (cursor != null) {
             if (cursor.moveToFirst()) {
+                mContentResolver.unregisterContentObserver(simChangeObserver);
                 int count = cursor.getCount();
 
                 for (int i = 0; i < count; ++i) {
+                    if (!mIsDeleteAll || cursor.isClosed()) {
+                        break;
+                    }
+                    cursor.moveToPosition(i);
                     deleteFromSim(cursor);
-                    cursor.moveToNext();
                 }
+                runOnUiThread(new Runnable() {
+                    @Override
+                    public void run() {
+                        refreshMessageList();
+                        registerSimChangeObserver();
+                    }
+                });
             }
         }
+        mIsDeleteAll = false;
     }
 
     @Override
